requirements.txt
google-adk>=0.1.0
gradio>=4.44.0
python-dotenv>=1.0.0
pymongo>=4.6.0
